# Optimize metrics collection