--- a/src/GraphQL.ApiTests/GraphQL.approved.txt
+++ b/src/GraphQL.ApiTests/GraphQL.approved.txt
@@ -944,7 +944,9 @@ namespace GraphQL.Instrumentation
     }
     public class Metrics
     {
+        public static readonly GraphQL.Instrumentation.Metrics Disabled;
         public Metrics(bool enabled = true) { }
+        public bool Enabled { get; }
         public GraphQL.Instrumentation.PerfRecord[] Finish() { }
         public GraphQL.Instrumentation.Metrics SetOperationName(string name) { }
         public GraphQL.Instrumentation.Metrics Start(string operationName) { }