--- a/src/GraphQL.ApiTests/GraphQL.approved.txt
+++ b/src/GraphQL.ApiTests/GraphQL.approved.txt
@@ -946,7 +946,7 @@ namespace GraphQL.Instrumentation
     {
         public Metrics(bool enabled = true) { }
         public bool Enabled { get; }
-        public static GraphQL.Instrumentation.Metrics Disabled { get; }
+        public static GraphQL.Instrumentation.Metrics None { get; }
         public GraphQL.Instrumentation.PerfRecord[] Finish() { }
         public GraphQL.Instrumentation.Metrics SetOperationName(string name) { }
         public GraphQL.Instrumentation.Metrics Start(string operationName) { }