--- a/src/GraphQL/Execution/ExecutionOptions.cs
+++ b/src/GraphQL/Execution/ExecutionOptions.cs
@@ -61,7 +61,7 @@ namespace GraphQL
         public List<IDocumentExecutionListener> Listeners { get; } = new List<IDocumentExecutionListener>();
 
         /// <summary>This setting essentially allows Apollo Tracing. Disabling will increase performance.</summary>
-        public bool EnableMetrics { get; set; } = true;
+        public bool EnableMetrics { get; set; }
 
         /// <summary>When false, captures unhandled exceptions and returns them within <see cref="ExecutionResult.Errors">ExecutionResult.Errors</see></summary>
         public bool ThrowOnUnhandledException { get; set; }