--- a/src/GraphQL/Execution/ExecutionStrategy.cs
+++ b/src/GraphQL/Execution/ExecutionStrategy.cs
@@ -133,7 +133,7 @@ namespace GraphQL.Execution
             var arrayItems = (data is ICollection collection)
                 ? new List<ExecutionNode>(collection.Count)
                 : new List<ExecutionNode>();
-            
+
             if (data is IList list)
             {
                 for (int i=0; i<list.Count; ++i)