--- a/src/GraphQL/Instrumentation/Metrics.cs
+++ b/src/GraphQL/Instrumentation/Metrics.cs
@@ -9,7 +9,6 @@ namespace GraphQL.Instrumentation
     /// </summary>
     public class Metrics
     {
-        private readonly bool _enabled;
         private ValueStopwatch _stopwatch;
         private readonly List<PerfRecord> _records;
         private PerfRecord _main;
@@ -22,7 +21,7 @@ namespace GraphQL.Instrumentation
         /// <param name="enabled">Indicates if metrics should be recorded for this execution.</param>
         public Metrics(bool enabled = true)
         {
-            _enabled = enabled;
+            Enabled = enabled;
             if (enabled)
                 _records = new List<PerfRecord>();
         }
@@ -30,7 +29,7 @@ namespace GraphQL.Instrumentation
         /// <summary>
         /// Shows whether metrics collection is enabled.
         /// </summary>
-        public bool Enabled => _enabled;
+        public bool Enabled { get; }
 
         /// <summary>
         /// Logs the start of the execution.
@@ -38,7 +37,7 @@ namespace GraphQL.Instrumentation
         /// <param name="operationName">The name of the GraphQL operation.</param>
         public Metrics Start(string operationName)
         {
-            if (_enabled)
+            if (Enabled)
             {
                 if (_main != null)
                     throw new InvalidOperationException("Metrics.Start has already been called");
@@ -56,7 +55,7 @@ namespace GraphQL.Instrumentation
         /// </summary>
         public Metrics SetOperationName(string name)
         {
-            if (_enabled && _main != null)
+            if (Enabled && _main != null)
                 _main.Subject = name;
 
             return this;
@@ -67,7 +66,7 @@ namespace GraphQL.Instrumentation
         /// </summary>
         public Marker Subject(string category, string subject, Dictionary<string, object> metadata = null)
         {
-            if (!_enabled)
+            if (!Enabled)
                 return Marker.Empty;
 
             if (_main == null)
@@ -84,7 +83,7 @@ namespace GraphQL.Instrumentation
         /// </summary>
         public PerfRecord[] Finish()
         {
-            if (!_enabled)
+            if (!Enabled)
                 return null;
 
             _main?.MarkEnd(_stopwatch.Elapsed.TotalMilliseconds);