``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.14393.3986 (1607/AnniversaryUpdate/Redstone1)
Intel Core i7-7700 CPU 3.60GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
Frequency=3515611 Hz, Resolution=284.4456 ns, Timer=TSC
.NET Core SDK=5.0.102
  [Host]     : .NET Core 3.1.11 (CoreCLR 4.700.20.56602, CoreFX 4.700.20.56604), X64 RyuJIT
  DefaultJob : .NET Core 3.1.11 (CoreCLR 4.700.20.56602, CoreFX 4.700.20.56604), X64 RyuJIT


```
before

|        Method | UseCaching |         Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |----------- |-------------:|----------:|----------:|---------:|--------:|------:|----------:|
| **Introspection** |      **False** | **1,112.457 μs** | **7.9694 μs** | **6.6548 μs** | **107.4219** | **33.2031** |     **-** | **487.09 KB** |
|          Hero |      False |    19.294 μs | 0.1714 μs | 0.1520 μs |   2.1057 |       - |     - |    8.7 KB |
| **Introspection** |       **True** |   **842.875 μs** | **6.8388 μs** | **5.7107 μs** | **106.4453** | **35.1563** |     **-** | **447.41 KB** |
|          Hero |       True |     3.722 μs | 0.0520 μs | 0.0487 μs |   0.9842 |       - |     - |   4.03 KB |

after

|        Method | UseCaching |         Mean |      Error |     StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |----------- |-------------:|-----------:|-----------:|---------:|--------:|------:|----------:|
| **Introspection** |      **False** | **1,112.892 μs** | **14.7801 μs** | **13.1022 μs** | **107.4219** | **33.2031** |     **-** | **486.28 KB** |
|          Hero |      False |    17.432 μs |  0.3071 μs |  0.2722 μs |   1.9226 |       - |     - |   7.89 KB |
| **Introspection** |       **True** |   **828.352 μs** | **14.5423 μs** | **13.6029 μs** | **104.4922** | **31.2500** |     **-** | **446.59 KB** |
|          Hero |       True |     3.211 μs |  0.0628 μs |  0.0588 μs |   0.7858 |       - |     - |   3.22 KB |


The results show improvement, but they are strange. A short `hero` request saves more memory than `introspection`.