@Shane32 2021-01-24
I don’t like the fact that there’s two properties on Metrics with similar names - an instance property called Enabled and a static method called Disabled. Any better ideas for that? Changes are fine otherwise.